# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI Boilerplate",
    "summary": "A modular FastAPI starter with a plugin system",
    "description": "\n    # FastAPI Boilerplate\n\n    A modern FastAPI starter with:\n\n    * Vertical-slice modules and a clean infrastructure layer\n    * Session-based auth with OAuth providers\n    * Swappable cache, queue, and rate-limit backends\n    * SQLAdmin admin UI\n    ",
    "contact": {
      "name": "Benav Labs",
      "url": "https://github.com/benavlabs/FastAPI-boilerplate",
      "email": "contact@benav.io"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.18.0"
  },
  "paths": {
...
    "/api/v1/api-keys/": {
      "post": {
        "tags": [
          "API Keys",
          "API Keys"
        ],
        "summary": "Create API Key",
        "description": "Creates a new API key for the authenticated user.\n\n    API keys enable programmatic access to the API and are useful\n    for building developer-facing products and integrations.\n\n    - **name**: Human-readable name for the API key\n    - **permissions**: Permission settings for the key\n    - **usage_limits**: Usage limits specific to this key\n    - **expires_at**: Optional expiration date\n\n    ⚠️ **Important**: The full API key is only shown once during creation.\n    Store it securely as it cannot be retrieved again.",
        "operationId": "create_api_key_api_v1_api_keys__post",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "name": "csrf_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Csrf Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/APIKeyCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API key created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Api Key Api V1 Api Keys  Post"
                }
              }
            }
          },
          "400": {
            "description": "Invalid API key data"
          },
          "401": {
            "description": "Authentication required"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "API Keys",
          "API Keys"
        ],
        "summary": "Get User Api Keys",
        "description": "Retrieves all API keys for the authenticated user.\n\n    - **active_only**: Whether to return only active keys (default: true)\n    - **page**: Page number (default: 1)\n    - **items_per_page**: Items per page (default: 50)\n\n    Returns keys sorted by creation date (newest first).\n    For security, only the key prefix is shown, not the full key.",
        "operationId": "get_user_api_keys_api_v1_api_keys__get",
        "parameters": [
          {
            "name": "active_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Return only active keys",
              "default": true,
              "title": "Active Only"
            },
            "description": "Return only active keys"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "items_per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 50,
              "title": "Items Per Page"
            },
            "description": "Items per page"
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "name": "csrf_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Csrf Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API keys retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedData_APIKeyRead_"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/api-keys/{key_id}": {
      "get": {
        "tags": [
          "API Keys",
          "API Keys"
        ],
        "summary": "Get API Key Details",
        "description": "Retrieves details for a specific API key.\n\n    - **key_id**: ID of the API key to retrieve\n\n    Users can only access their own API keys.\n    Returns comprehensive key information including usage limits and permissions.",
        "operationId": "get_api_key_api_v1_api_keys__key_id__get",
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "API key ID",
              "title": "Key Id"
            },
            "description": "API key ID"
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "name": "csrf_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Csrf Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API key details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Api Key Api V1 Api Keys  Key Id  Get"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied to this API key"
          },
          "404": {
            "description": "API key not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "API Keys",
          "API Keys"
        ],
        "summary": "Update API Key",
        "description": "Updates an existing API key.\n\n    - **key_id**: ID of the API key to update\n\n    Allows updating name, permissions, usage limits, active status, and expiration.\n    Users can only update their own API keys.",
        "operationId": "update_api_key_api_v1_api_keys__key_id__patch",
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "API key ID",
              "title": "Key Id"
            },
            "description": "API key ID"
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "name": "csrf_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Csrf Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/APIKeyUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "API key updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Api Key Api V1 Api Keys  Key Id  Patch"
                }
              }
            }
          },
          "400": {
            "description": "Invalid update data"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied to this API key"
          },
          "404": {
            "description": "API key not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "API Keys",
          "API Keys"
        ],
        "summary": "Delete API Key",
        "description": "Deletes (deactivates) an API key.\n\n    - **key_id**: ID of the API key to delete\n\n    This operation deactivates the key rather than permanently deleting it\n    to maintain usage history and audit trails.\n\n    Users can only delete their own API keys.",
        "operationId": "delete_api_key_api_v1_api_keys__key_id__delete",
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "API key ID",
              "title": "Key Id"
            },
            "description": "API key ID"
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "name": "csrf_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Csrf Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API key deleted successfully",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied to this API key"
          },
          "404": {
            "description": "API key not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/api-keys/{key_id}/usage": {
      "get": {
        "tags": [
          "API Keys",
          "API Keys"
        ],
        "summary": "Get Key Usage",
        "description": "Retrieves usage history for a specific API key.\n\n    - **key_id**: ID of the API key\n    - **page**: Page number (default: 1)\n    - **items_per_page**: Items per page (default: 100)\n\n    Returns usage records in reverse chronological order (newest first).\n    Includes details like endpoints used, response times, costs, and errors.",
        "operationId": "get_key_usage_api_v1_api_keys__key_id__usage_get",
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "API key ID",
              "title": "Key Id"
            },
            "description": "API key ID"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "items_per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Items per page",
              "default": 100,
              "title": "Items Per Page"
            },
            "description": "Items per page"
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "name": "csrf_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Csrf Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usage history retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedData_KeyUsageRead_"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Access denied to this API key"
          },
          "404": {
            "description": "API key not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIKeyCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "description": "Human-readable name for the API key"
          },
          "permissions": {
            "additionalProperties": true,
            "type": "object",
            "title": "Permissions",
            "description": "Permission settings"
          },
          "usage_limits": {
            "additionalProperties": true,
            "type": "object",
            "title": "Usage Limits",
            "description": "Usage limits per key"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At",
            "description": "Key expiration timestamp"
          },
          "key_metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key Metadata",
            "description": "Additional key metadata"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "APIKeyCreate",
        "description": "Schema for creating a new API key."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}